:recycle: Move BFT to node
