+++ framework/src/node/bft/bft_votes.ts
@@ -12,7 +12,7 @@
  * Removal or modification of this copyright notice is prohibited.
  */
 
-import { BlockHeader } from '../../node/state_machine';
+import { BlockHeader } from '../state_machine';
 import { BFTParametersCache } from './bft_params';
 import { BFTVotes } from './schemas';
 import { getBlockBFTProperties } from './utils';
