+++ framework/src/node/bft/errors.ts
