+++ framework/src/node/bft/index.ts
