+++ framework/src/node/bft/schemas.ts
