+++ framework/src/node/bft/types.ts
